move assert usage to go117+
