+++ go.mod
@@ -8,6 +8,6 @@ require (
 	github.com/golang-sql/civil v0.0.0-20220223132316-b832511892a9
 	github.com/golang-sql/sqlexp v0.1.0
 	github.com/jcmturner/gokrb5/v8 v8.4.4
+	github.com/stretchr/testify v1.8.1
 	golang.org/x/crypto v0.9.0
 )
-
