+++ tds_test.go
@@ -4,13 +4,11 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"crypto/rand"
 	"database/sql"
 	"encoding/binary"
 	"encoding/hex"
 	"fmt"
 	"io"
-	"math/big"
 	"net"
 	"net/url"
 	"os"
@@ -23,7 +21,6 @@ import (
 	"unicode/utf16"
 
 	"github.com/microsoft/go-mssqldb/msdsn"
-	"github.com/stretchr/testify/assert"
 )
 
 type MockTransport struct {
@@ -1026,48 +1023,4 @@ func TestDialSqlConnectionCustomDialer(t *testing.T) {
 	if err == nil {
 		t.Error(fmt.Errorf("dialer should not be used to resolve dns if not a host dialer"))
 	}
-}
-
-func TestChangePassword(t *testing.T) {
-	conn, logger := open(t)
-	defer conn.Close()
-	defer logger.StopLogging()
-	login, pwd := createLogin(t, conn)
-	defer dropLogin(t, conn, login)
-	p, err := msdsn.Parse(makeConnStr(t).String())
-	assert.NoError(t, err, "Parse failed")
-	p.ChangePassword = "Change" + pwd
-	p.User = login
-	p.Password = pwd
-	p.Parameters[msdsn.UserId] = p.User
-	p.Parameters[msdsn.Password] = p.Password
-	tl := testLogger{t: t}
-	defer tl.StopLogging()
-	c, err := connect(context.Background(), &Connector{params: p}, optionalLogger{loggerAdapter{&tl}}, p)
-	if assert.NoError(t, err, "Login with new login failed") {
-		c.buf.transport.Close()
-
-		p.Password = p.ChangePassword
-		p.ChangePassword = ""
-		c, err = connect(context.Background(), &Connector{params: p}, optionalLogger{loggerAdapter{&tl}}, p)
-		if assert.NoError(t, err, "Login with new password failed") {
-			c.buf.transport.Close()
-		}
-	}
-
-}
-
-func createLogin(t *testing.T, conn *sql.DB) (login string, password string) {
-	t.Helper()
-	suffix, _ := rand.Int(rand.Reader, big.NewInt(10000))
-	login = fmt.Sprintf("mssqlLogin%d", suffix.Int64())
-	password = fmt.Sprintf("mssqlPwd!%d", suffix.Int64())
-	_, err := conn.Exec(fmt.Sprintf("CREATE LOGIN [%s] WITH PASSWORD = '%s', CHECK_POLICY=OFF", login, password))
-	assert.NoError(t, err, "create login failed")
-	return
-}
-
-func dropLogin(t *testing.T, conn *sql.DB, login string) {
-	t.Helper()
-	_, _ = conn.Exec(fmt.Sprintf("DROP LOGIN [%s]", login))
 }
